remote_settings: fix fetching config route
The config route setting has no effect due to a bug where pybrake doesn't read
that value prior to making a GET request to the notifier config server. The fix
is to call `config_route()` on every notifier config GET call.

Additionally, we are adding a logic that prevents remote settings to crash the
background thread in case `config_route` is a bad value (HTTP lib would raise an
error). When an error happens such as 403 Forbidden, we make a 2nd request to
the old config route, which was known to work.